Refactor map layer utilities and constants structure
Moved map layer utility functions, hooks, and constants to centralized locations under 'features/map/utils', 'features/map/hooks', and 'features/map/constants'. Updated all relevant imports and fixed a typo in the Threshold selector filename. This improves code organization and maintainability.
