+++ src/app/CreateScenario.jsx
@@ -29,7 +29,7 @@ import {
 } from 'features/scenario/hooks/create-scenario-forms';
 import { LoadingOutlined } from '@ant-design/icons';
 import { useCameraForBounds } from 'features/map/hooks';
-import { calcBoundsAndCenter } from 'features/map/components/Map/utils';
+import { calcBoundsAndCenter } from 'features/map/utils';
 
 import routes from 'constants/routes.json';
 import { useProjectStore } from 'features/project/stores/projectStore';
